chore(suite): rename ChangeFee components index > name
