+++ packages/suite/src/components/suite/modals/TransactionDetail/components/ChangeFee/components/GreyCard.tsx
@@ -12,6 +12,6 @@ const Wrapper = styled(Card)`
     margin-top: 12px;
 `;
 
-const GreyCard = (props: { children?: React.ReactNode }) => <Wrapper>{props.children}</Wrapper>;
-
-export default GreyCard;
+export const GreyCard = (props: { children?: React.ReactNode }) => (
+    <Wrapper>{props.children}</Wrapper>
+);
