+++ packages/suite/src/components/suite/modals/TransactionDetail/components/ChangeFee/components/RbfFees.tsx
@@ -3,7 +3,7 @@ import { Fees } from 'src/components/wallet/Fees';
 import { useRbfContext } from 'src/hooks/wallet/useRbfForm';
 
 // wrapper for shareable Fees component
-const RbfFees = () => {
+export const RbfFees = () => {
     const {
         formState: { errors },
         register,
@@ -32,5 +32,3 @@ const RbfFees = () => {
         />
     );
 };
-
-export default RbfFees;
